requirements.txt
biopython==1.83
contourpy==1.2.1
cycler==0.12.1
fonttools==4.53.1
joblib==1.4.2
kiwisolver==1.4.5
matplotlib==3.9.1
numpy==1.26.4
packaging==24.1
pandas==2.2.2
pillow==10.4.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
scikit-learn==1.5.0
scipy==1.13.1
six==1.16.0
threadpoolctl==3.5.0
tzdata==2024.1
